+++ device/udp/udpGetbuf.c
@@ -5,6 +5,7 @@
 /* Embedded Xinu, Copyright (C) 2009.  All rights reserved. */
 
 #include <stddef.h>
+#include <stdint.h>
 #include <bufpool.h>
 #include <network.h>
 #include <stdlib.h>
@@ -22,7 +23,7 @@ struct udpPkt *udpGetbuf(struct udp *udpptr)
     struct udpPkt *udppkt = NULL;
 
     udppkt = bufget(udpptr->inPool);
-    if (SYSERR == (int)udppkt)
+    if (SYSERR == (uintptr_t)udppkt)
     {
         return (struct udpPkt *)SYSERR;
     }
