+++ include/safemem.h
@@ -10,14 +10,15 @@
 #define _SAFEMEM_H_
 
 #include <stddef.h>
+#include <stdint.h>
 #include <conf.h>
 
 #define PAGE_SIZE 4096
 
 /** メモリアドレスを上位ページ境界に丸める */
-#define roundpage(x) ((4095 + (uint)(x)) & ~0x0FFF)
+#define roundpage(x) ((4095 + (uint64_t)(x)) & ~0x0FFF)
 /** メモリアドレスを下位ページ境界に切り捨てる */
-#define truncpage(x) ((uint)(x) & ~0x0FFF)
+#define truncpage(x) ((uint64_t)(x) & ~0x0FFF)
 
 /* Region allocator */
 
@@ -29,7 +30,7 @@ struct memregion
     struct memregion *prev;     /**< pointer to previous region in list */
     struct memregion *next;     /**< pointer to next region in list */
     void *start;                /**< Starting address (page aligned) */
-    uint length;                /**< Size of region */
+    uint32_t length;                /**< Size of region */
     tid_typ thread_id;          /**< Holding thread identifier */
 };
 
@@ -38,12 +39,12 @@ extern struct memregion *regalloclist; /**< List of allocated regions */
 extern struct memregion *regtab;       /**< Array of regions */
 
 /* Prototypes for memory region allocator */
-void memRegionInit(void *, uint);
+void memRegionInit(void *, uint32_t);
 void memRegionClear(struct memregion *);
 void memRegionInsert(struct memregion *, struct memregion **);
 void memRegionRemove(struct memregion *, struct memregion **);
-struct memregion *memRegionAlloc(uint);
-struct memregion *memRegionSplit(struct memregion *, uint length);
+struct memregion *memRegionAlloc(uint32_t);
+struct memregion *memRegionSplit(struct memregion *, uint32_t length);
 struct memregion *memRegionValid(void *);
 void memRegionTransfer(void *, tid_typ);
 void memRegionReclaim(tid_typ);
@@ -67,14 +68,14 @@ struct pgtblent
 };
 
 extern struct pgtblent *pgtbl;      /**< system page table */
-extern uint pgtbl_nents;            /**< number of pages in page table */
+extern uint32_t pgtbl_nents;            /**< number of pages in page table */
 
 /* Prototypes for memory protection functions */
 void safeInit(void);
-int safeMap(void *, short);
-int safeMapRange(void *, uint, short);
+int safeMap(void *, int16_t);
+int safeMapRange(void *, uint32_t, int16_t);
 int safeUnmap(void *);
-int safeUnmapRange(void *, uint);
+int safeUnmapRange(void *, uint32_t);
 void safeKmapInit(void);
 
 #endif                          /* _SAFEMEM_H_ */
