+++ include/stdio.h
@@ -6,6 +6,7 @@
 #ifndef _STDIO_H_
 #define _STDIO_H_
 
+#include <stdint.h>
 #include <compiler.h>
 #include <stdarg.h>
 #include <thread.h>  /* For thrtab and thrcurrent. */
@@ -33,7 +34,7 @@ extern unsigned int getcpuid(void);
 
 /* Formatted input  */
 int _doscan(const char *fmt, va_list ap,
-            int (*getch) (int, int), int (*ungetch) (int, int),
+            int (*getch) (int, uintptr_t), int (*ungetch) (int, uintptr_t),
             int arg1, int arg2);
 
 int fscanf(int dev, const char *format, ...);
@@ -47,7 +48,7 @@ int sscanf(const char *str, const char *format, ...);
 
 /* Formatted output  */
 int _doprnt(const char *format, va_list,
-	    int (*putc_func)(int, int), int putc_arg);
+	    int (*putc_func)(int, uintptr_t), uintptr_t putc_arg);
 
 int fprintf(int dev, const char *format, ...) __printf_format(2, 3);
 int printf(const char *format, ...) __printf_format(1, 2);
@@ -56,7 +57,7 @@ int sprintf(char *str, const char *format, ...) __printf_format(2, 3);
 /* Character and string input and output  */
 int fgetc(int dev);
 char *fgets(char *s, int n, int dev);
-int fputc(int c, int dev);
+int fputc(int c, uintptr_t dev);
 int fputs(const char *s, int dev);
 
 /** @ingroup libxc */
