+++ lib/libxc/doscan.c
@@ -20,14 +20,14 @@ enum integer_size {
 
 static int scan_string(char *ptr, int type, uint maxlen,
                        const uchar *stopchar_tab,
-                       int (*getch) (int, int), int (*ungetch) (int, int),
+                       int (*getch) (int, uintptr_t), int (*ungetch) (int, uintptr_t),
                        int arg1, int arg2, bool *eofptr);
 
 static int scan_number_or_string(void *ptr, uchar type, uint maxlen,
                                  enum integer_size size,
                                  const uchar *stopchar_tab,
-                                 int (*getch) (int, int),
-                                 int (*ungetch) (int, int),
+                                 int (*getch) (int, uintptr_t),
+                                 int (*ungetch) (int, uintptr_t),
                                  int arg1, int arg2, bool *eofptr);
 
 static const uchar *build_stopchar_tab(const uchar *ufmt, uchar *stopchar_tab);
@@ -81,7 +81,7 @@ static const uchar *build_stopchar_tab(const uchar *ufmt, uchar *stopchar_tab);
  *      returned.
  */
 int _doscan(const char *fmt, va_list ap,
-            int (*getch) (int, int), int (*ungetch) (int, int),
+            int (*getch) (int, uintptr_t), int (*ungetch) (int, uintptr_t),
             int arg1, int arg2)
 {
     int nmatch;
@@ -235,7 +235,7 @@ int _doscan(const char *fmt, va_list ap,
 }
 
 static int scan_string(char *ptr, int type, uint maxlen, const uchar *stopchar_tab,
-                       int (*getch) (int, int), int (*ungetch) (int, int),
+                       int (*getch) (int, uintptr_t), int (*ungetch) (int, uintptr_t),
                        int arg1, int arg2, bool *eofptr)
 {
     uint len;
@@ -290,8 +290,8 @@ static int scan_string(char *ptr, int type, uint maxlen, const uchar *stopchar_t
 static int scan_number_or_string(void *ptr, uchar type, uint maxlen,
                                  enum integer_size size,
                                  const uchar *stopchar_tab,
-                                 int (*getch) (int, int),
-                                 int (*ungetch) (int, int),
+                                 int (*getch) (int, uintptr_t),
+                                 int (*ungetch) (int, uintptr_t),
                                  int arg1, int arg2, bool *eofptr)
 {
     int c = EOF;
