+++ shell/xsh_snoop.c
@@ -5,6 +5,7 @@
 /* Embedded Xinu, Copyright (C) 2009.  All rights reserved. */
 
 #include <stddef.h>
+#include <stdint.h>
 #include <conf.h>
 #include <ipv4.h>
 #include <shell.h>
@@ -77,7 +78,7 @@ static thread snoop(struct snoop *cap, uint count, char dump,
     while (forever || count > 0)
     {
         pkt = snoopRead(cap);
-        if ((SYSERR == (int)pkt) || (NULL == pkt))
+        if ((SYSERR == (uintptr_t)pkt) || (NULL == pkt))
         {
             continue;
         }
